Big update. Changed the main component from JTextArea to JTextPane
Changed the main component from JTextArea to JTextPane in order to match the dynamic features that the project will eventually include. Converted the necessary methods and interfaces to match the new component.
Added custom methods to MyJTextArea class to ease the use of the JTextPane component that the class modifies.
Removed word wrapping as it is default in JTextPane.
Added basic dynamic format features - bold and italic text.
Updated menu & key bindings (Ctrl+b, Ctrl+i) for the bold/italic modification.